requirements.txt
ipwhois
colorlog
scapy
discord
requests
python-nmap
appdirs==1.4.3
backports.functools-lru-cache==1.6.1
beautifulsoup4==4.9.0
configparser==4.0.2
contextlib2==0.6.0.post1
distlib==0.3.0
dnspython==1.16.0
filelock==3.0.12
importlib-metadata==1.6.0
importlib-resources==1.4.0
netaddr==0.7.19
pathlib2==2.3.5
scandir==1.10.0
singledispatch==3.4.0.3
six==1.14.0
soupsieve==1.9.5
typing==3.7.4.1
virtualenv==20.0.17
zipp==1.2.0
